more windows compiler debugging
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@2475 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/libs/orcaice/detail/localstatus.cpp
@@ -152,7 +152,7 @@ LocalStatus::process()
     for ( it=subsystems_.begin(); it!=subsystems_.end(); ++it ) 
     {
         IceUtil::Time timeSinceLastHeartbeat = now-it->second.lastHeartbeatTime;
-        double secSinceLastHeartbeat = timeSinceLastHeartbeat.toSeconds();
+        double secSinceLastHeartbeat = (double)timeSinceLastHeartbeat.toSeconds();
         if ( secSinceLastHeartbeat > it->second.maxHeartbeatInterval )
         {
             // Oops, this subsystem appears to be dead...
